+++ cliche/commands/__init__.py
@@ -8,6 +8,7 @@ from .code import code
 from .process import kill
 from .roast import roastme
 from .scrape import scrape
+from .search import search
 from .server import servers
 from .system import system
 from .view import view
@@ -22,6 +23,7 @@ __all__ = [
     'kill',
     'roastme',
     'scrape',
+    'search',
     'servers',
     'system',
     'view',